requirements.txt
streamlit==1.35.0
streamlit-drawable-canvas==0.9.3
opencv-python-headless==4.9.0.80
pillow==10.3.0
numpy==1.26.4